Reorganizing code for clarity
