+++ functions/src/consts.ts
@@ -1,0 +1,1 @@
+export const EXPIRATION_TIMESTAMP_KEY = "expirationTimestamp";
